requirements/base.txt
#Created by "reqpy"
#https://github.com/gopar/reqpy
Django==1.9.7
djangorestframework==3.3.3
Markdown==2.6.6
